Make ISO C99 -pedantic _POSIX_C_SOURCE 200112L compatible

+++ include/md5.h
@@ -13,16 +13,16 @@
 #include <stddef.h>
 
 /****************************** MACROS ******************************/
-#define MD5_BLOCK_SIZE 16               // MD5 outputs a 16 byte digest
+#define MD5_BLOCK_SIZE 16               /* MD5 outputs a 16 byte digest */
 
 /**************************** DATA TYPES ****************************/
-typedef unsigned char BYTE;             // 8-bit byte
-typedef unsigned int  WORD;             // 32-bit word, change to "long" for 16-bit machines
+typedef unsigned char BYTE;             /* 8-bit byte */
+typedef unsigned int  WORD;             /* 32-bit word, change to "long" for 16-bit machines */
 
 typedef struct {
    BYTE data[64];
    WORD datalen;
-   unsigned long long bitlen;
+   unsigned long int bitlen;
    WORD state[4];
 } MD5_CTX;
 
@@ -31,4 +31,4 @@ void md5_init(MD5_CTX *ctx);
 void md5_update(MD5_CTX *ctx, const BYTE data[], size_t len);
 void md5_final(MD5_CTX *ctx, BYTE hash[]);
 
-#endif   // MD5_H
+#endif   /* MD5_H */
